requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.21.0
spacy>=3.0.0
sentence-transformers>=2.2.0
torch>=1.7.0
